requirements.txt
flask==3.0.0
torch>=2.5.0
torchvision>=0.20.0
open-clip-torch==2.23.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
werkzeug==3.0.1
gunicorn==21.2.0
